fixup swapchecker tests
