+++ src/components/SwapChecker.tsx
@@ -1,20 +1,110 @@
 import log from "loglevel";
 import { createEffect, createSignal } from "solid-js";
 
-import { swap, swaps } from "../signals";
-import { fetcher, getApiUrl, setSwapStatusAndClaim } from "../utils/helper";
-import { swapStatusFinal } from "../utils/swapStatus";
-
-const swapCheckInterval = 3000;
-
-let activeStreamId = undefined;
-let activeSwapStream = undefined;
+import { RBTC } from "../consts";
+import {
+    setFailureReason,
+    setSwapStatus,
+    setSwapStatusTransaction,
+    setTimeoutBlockheight,
+    setTimeoutEta,
+    swap,
+    swaps,
+} from "../signals";
+import { claim, fetcher, getApiUrl } from "../utils/helper";
+import {
+    swapStatusFinal,
+    swapStatusPending,
+    updateSwapStatus,
+} from "../utils/swapStatus";
 
 export const [checkInterval, setCheckInterval] = createSignal<
     NodeJS.Timer | undefined
 >(undefined);
 
 export const SwapChecker = () => {
+    const swapCheckInterval = 3000;
+
+    let activeStreamId = undefined;
+    let activeSwapStream = undefined;
+
+    const checkForFailed = (swap: any, data: any) => {
+        if (
+            data.status == "transaction.lockupFailed" ||
+            data.status == "invoice.failedToPay"
+        ) {
+            const id = swap.id;
+
+            fetcher(
+                "/getswaptransaction",
+                swap.asset,
+                (data: any) => {
+                    if (swap.asset !== RBTC && !data.transactionHex) {
+                        log.error("no mempool tx found");
+                    }
+                    if (!data.timeoutEta) {
+                        log.error("no timeout eta");
+                    }
+                    if (!data.timeoutBlockHeight) {
+                        log.error("no timeout blockheight");
+                    }
+                    const timestamp = data.timeoutEta * 1000;
+                    const eta = new Date(timestamp);
+                    log.debug(
+                        "Timeout ETA: \n " + eta.toLocaleString(),
+                        timestamp,
+                    );
+                    setTimeoutEta(timestamp);
+                    setTimeoutBlockheight(data.timeoutBlockHeight);
+                },
+                {
+                    id,
+                },
+            );
+        }
+    };
+
+    const setSwapStatusAndClaim = (data: any, activeSwap: any) => {
+        const currentSwap = swaps().find((s) => activeSwap.id === s.id);
+        if (swap() && swap().id === currentSwap.id) {
+            setSwapStatus(data.status);
+        }
+
+        if (data.transaction) setSwapStatusTransaction(data.transaction);
+        if (data.status) updateSwapStatus(currentSwap.id, data.status);
+
+        if (
+            currentSwap.claimTx === undefined &&
+            data.transaction !== undefined &&
+            (data.status === swapStatusPending.TransactionConfirmed ||
+                data.status === swapStatusPending.TransactionMempool)
+        ) {
+            claim(currentSwap);
+        }
+        checkForFailed(currentSwap, data);
+        if (data.failureReason) setFailureReason(data.failureReason);
+    };
+
+    const runSwapCheck = async () => {
+        const swapsToCheck = swaps()
+            .filter((s) => !swapStatusFinal.includes(s.status))
+            .filter((s) => s.id !== swap()?.id);
+
+        for (const swap of swapsToCheck) {
+            await new Promise<void>((resolve) => {
+                fetcher(
+                    "/swapstatus",
+                    swap.asset,
+                    (data: any) => {
+                        setSwapStatusAndClaim(data, swap);
+                        resolve();
+                    },
+                    { id: swap.id },
+                );
+            });
+        }
+    };
+
     createEffect(() => {
         const activeSwap = swap();
         if (swap()?.id === activeStreamId) {
@@ -69,26 +159,6 @@ export const SwapChecker = () => {
     );
 
     return "";
-};
-
-const runSwapCheck = async () => {
-    const swapsToCheck = swaps()
-        .filter((s) => !swapStatusFinal.includes(s.status))
-        .filter((s) => s.id !== swap()?.id);
-
-    for (const swap of swapsToCheck) {
-        await new Promise<void>((resolve) => {
-            fetcher(
-                "/swapstatus",
-                swap.asset,
-                (data: any) => {
-                    setSwapStatusAndClaim(data, swap);
-                    resolve();
-                },
-                { id: swap.id },
-            );
-        });
-    }
 };
 
 const handleStream = (streamUrl: string, cb: (data: any) => void) => {
